+++ src/main/java/hr/dlatecki/algorithms/gen_alg/exceptions/TooFewParentsException.java
@@ -1,0 +1,34 @@
+// Copyright (C) Domagoj Latečki
+// You may use, distribute and modify this code under the terms of the MIT license.
+package hr.dlatecki.algorithms.gen_alg.exceptions;
+
+/**
+ * Thrown to indicate that there are too few parent chromosomes provided.
+ * 
+ * @author Domagoj Latečki
+ * @version 1.0
+ * @since 1.8
+ */
+public class TooFewParentsException extends RuntimeException {
+    
+    /**
+     * Serial version UID.
+     */
+    private static final long serialVersionUID = 258653111631295365L;
+    
+    /**
+     * Constructs the <code>TooFewParentsException</code> with no detail message.
+     */
+    public TooFewParentsException() {
+        super();
+    }
+    
+    /**
+     * Constructs the <code>TooFewParentsException</code> with specified detail message.
+     * 
+     * @param message the detail message.
+     */
+    public TooFewParentsException(String message) {
+        super(message);
+    }
+}