+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Development history
 
+### v0.6.2
+Added TooFewParentsException. AbstractCrossoverOperator now has default behaviour for doCrossover(SortedSet<C>) method
+and defines an abstract method which must be overridden in derived classes to generate a child chromosome.
+
 ### v0.6.1
 Removed annotations package and all its contents. Added MIN\_SELECTION\_SIZE constant in ISelection operator interface.
 All chromosome implementations are now marked as final.